Use a new common modal to show the fixture contents using the store contents
