+++ web-app/src/constants/modals.ts
@@ -1,0 +1,9 @@
+import FutsalScoreModal from '@/components/futsal/FutsalScoreModal.vue';
+import TableTennisScoreModal from '@/components/table-tennis/TableTennisScoreModal.vue';
+import CarromBoardScoreModal from '@/components/carrom-board/CarromBoardScoreModal.vue';
+
+export const FUTSAL_SCORE_MODAL = FutsalScoreModal;
+
+export const TABLE_TENNIS_SCORE_MODAL = TableTennisScoreModal;
+
+export const CARROM_BOARD_SCORE_MODAL = CarromBoardScoreModal;
