+++ web-app/src/store/index.ts
@@ -1,6 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
+import scoreModal from './modules/scoreModal';
 import tournaments from './modules/tournaments';
 
 Vue.use(Vuex);
@@ -9,6 +10,7 @@ const debug = process.env.NODE_ENV !== 'production';
 
 export default new Vuex.Store({
   modules: {
+    scoreModal,
     tournaments
   },
   strict: debug
